# api/api.yaml
openapi: "3.0.3"
info:
  title: Go Simple ZFS Api
  version: "1.0"
servers:
  - url: https://api.server.test/v1
paths:
  /readyz:
    get:
      responses:
        "200":
          description: Service is ready
  /zfs/list:
    get:
      summary: List ZFS datasets
      operationId: listZfsDatasets
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ZFSDatasetType"
      responses:
        "200":
          description: A list of ZFS datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZFSDataset"
  /zfs/clone:
    post:
      summary: Clone a ZFS dataset
      operationId: cloneZfsDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZFSCloneReq"
      responses:
        "201":
          description: Dataset cloned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZFSDataset"
        "400":
          description: Bad request, invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  # /zfs/get:
  #   get:
  #     summary: Get ZFS datasets
  #     operationId: getZfsDatasets
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/ZFSDatasetGetReq"
  #     responses:
  #       "404":
  #         description: Dataset not found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "200":
  #         description: A list of ZFS datasets
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ZFSDatasetGetResp"
components:
  schemas:
    ZFSCloneReq:
      type: object
      required:
        - source
        - target
      properties:
        source:
          type: string
          description: Source dataset to clone from
        target:
          type: string
          description: Target dataset name for the clone
    ZFSDatasetType:
      type: string
      default: all
      enum:
        - filesystem
        - snapshot
        - volume
        - bookmark
        - all
    ZFSDatasetProp:
      type: object
      required:
        - name
        - value
      properties:
        name:
          type: string
          description: Name of the dataset property
        value:
          type: string
          description: Value of the dataset property
    ZFSDatasetGetReq:
      type: object
      required:
        - name
        - properties
      properties:
        name:
          type: string
          description: Name of dataset to retrieve
        properties:
          type: array
          items:
            type: string
          description: List of properties to retrieve for the dataset
    ZFSDatasetGetResp:
      type: object
      required:
        - properties
      properties:
        properties:
          type: array
          items:
            $ref: "#/components/schemas/ZFSDatasetProp"
    ZFSDataset:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        used:
          type: string
        available:
          type: string
        mountpoint:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message