# internal/models/models.gen.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package models

// Defines values for ZFSDatasetType.
const (
	All        ZFSDatasetType = "all"
	Bookmark   ZFSDatasetType = "bookmark"
	Filesystem ZFSDatasetType = "filesystem"
	Snapshot   ZFSDatasetType = "snapshot"
	Volume     ZFSDatasetType = "volume"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Error message
	Error string `json:"error"`
}

// ZFSCloneReq defines model for ZFSCloneReq.
type ZFSCloneReq struct {
	// Source Source dataset to clone from
	Source string `json:"source"`

	// Target Target dataset name for the clone
	Target string `json:"target"`
}

// ZFSDataset defines model for ZFSDataset.
type ZFSDataset struct {
	Available  *string `json:"available,omitempty"`
	Mountpoint *string `json:"mountpoint,omitempty"`
	Name       string  `json:"name"`
	Used       *string `json:"used,omitempty"`
}

// ZFSDatasetType defines model for ZFSDatasetType.
type ZFSDatasetType string

// ListZfsDatasetsParams defines parameters for ListZfsDatasets.
type ListZfsDatasetsParams struct {
	Type *ZFSDatasetType `form:"type,omitempty" json:"type,omitempty"`
}

// CloneZfsDatasetJSONRequestBody defines body for CloneZfsDataset for application/json ContentType.
type CloneZfsDatasetJSONRequestBody = ZFSCloneReq